+++ libs/utils/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @protolabsai/utils
 
+## 0.99.3
+
+### Patch Changes
+
+- Maintenance release.
+- Updated dependencies
+  - @protolabsai/types@0.99.3
+  - @protolabsai/platform@0.99.3
+
+## 0.99.2
+
+### Patch Changes
+
+- Maintenance release.
+- Updated dependencies
+  - @protolabsai/types@0.99.2
+  - @protolabsai/platform@0.99.2
+
+## 0.99.1
+
+### Patch Changes
+
+- Maintenance release.
+- Updated dependencies
+  - @protolabsai/types@0.99.1
+  - @protolabsai/platform@0.99.1
+
 ## 0.99.0
 
 ### Minor Changes